+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @sap-ux/fiori-elements-writer
 
+## 1.3.11
+
+### Patch Changes
+
+-   Updated dependencies [231e713]
+    -   @sap-ux/fiori-generator-shared@0.7.0
+
+## 1.3.10
+
+### Patch Changes
+
+-   f626d47: remove @sap/ux-specification from generated apps
+-   Updated dependencies [f626d47]
+    -   @sap-ux/ui5-application-writer@1.2.1
+
+## 1.3.9
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.11
+
+## 1.3.8
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.10
+-   @sap-ux/fiori-generator-shared@0.6.4
+-   @sap-ux/odata-service-writer@0.23.0
+-   @sap-ux/ui5-application-writer@1.2.0
+-   @sap-ux/ui5-test-writer@0.4.1
+
+## 1.3.7
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.9
+-   @sap-ux/fiori-generator-shared@0.6.3
+-   @sap-ux/odata-service-writer@0.23.0
+-   @sap-ux/ui5-application-writer@1.2.0
+-   @sap-ux/ui5-test-writer@0.4.1
+
+## 1.3.6
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.8
+-   @sap-ux/fiori-generator-shared@0.6.2
+-   @sap-ux/odata-service-writer@0.23.0
+-   @sap-ux/ui5-application-writer@1.2.0
+-   @sap-ux/ui5-test-writer@0.4.1
+
 ## 1.3.5
 
 ### Patch Changes