+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/inquirer-common
 
+## 0.4.8
+
+### Patch Changes
+
+-   02e4f29: add resource bundle for inquirer common in consumers
+
+## 0.4.7
+
+### Patch Changes
+
+-   d29b1a3: update ui5 default version logic
+
 ## 0.4.6
 
 ### Patch Changes