+++ packages/launch-config/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/launch-config
 
+## 0.7.6
+
+### Patch Changes
+
+-   81a15ab: Fix type error for empty launch.json file.
+
+## 0.7.5
+
+### Patch Changes
+
+-   Updated dependencies [eb38e5b]
+    -   @sap-ux/project-access@1.28.2
+
+## 0.7.4
+
+### Patch Changes
+
+-   Updated dependencies [64e037d]
+    -   @sap-ux/project-access@1.28.1
+
+## 0.7.3
+
+### Patch Changes
+
+-   Updated dependencies [15e6959]
+    -   @sap-ux/project-access@1.28.0
+
 ## 0.7.2
 
 ### Patch Changes